+++ irmaclient/legacy.go
@@ -558,3 +558,17 @@ func (kss *keyshareServer) registerPublicKey(client *Client, transport *irma.HTT
 
 	return result, nil
 }
+
+// removeKeysharePsFromProofUs fixes a difference in gabi between the old keyshare protocol and
+// the new one. In the old one, during issuance the client sends a proof of knowledge only of its
+// own keyshare to the issuer. In the new one, it sends a proof of knowledge of the full secret.
+// Therefore, the proofU contains a PoK over the full secret, while in case of the old keyshare
+// protocol, the issuer expects a PoK only of the user's keyshare. This method removes the
+// keyshare server's contribution for use in the old keyshare protocol.
+func (ks *keyshareSession) removeKeysharePsFromProofUs(proofs gabi.ProofList) {
+	for i, proof := range proofs {
+		if proofU, ok := proof.(*gabi.ProofU); ok {
+			proofU.RemoveKeyshareP(ks.builders[i].(*gabi.CredentialBuilder))
+		}
+	}
+}