+++ messages.go
@@ -3,6 +3,7 @@ package irma
 import (
 	"bytes"
 	"encoding/json"
+	"github.com/privacybydesign/gabi/big"
 	"net/url"
 	"regexp"
 	"strconv"
@@ -396,6 +397,53 @@ func (ppcm *ProofPCommitmentMap) MarshalJSON() ([]byte, error) {
 	}
 
 	return json.Marshal(encPPCM)
+}
+
+type PMap struct {
+	Ps map[PublicKeyIdentifier]*big.Int `json:"ps"`
+}
+
+func (pm *PMap) MarshalJSON() ([]byte, error) {
+	var encPM struct {
+		Ps map[string]*big.Int `json:"ps"`
+	}
+	encPM.Ps = make(map[string]*big.Int)
+
+	for pki, v := range pm.Ps {
+		pkiBytes, err := pki.MarshalText()
+		if err != nil {
+			return nil, err
+		}
+		encPM.Ps[string(pkiBytes)] = v
+	}
+
+	return json.Marshal(encPM)
+}
+
+type GetCommitmentsRequest struct {
+	Keys []PublicKeyIdentifier          `json:"keys"`
+	Hash gabi.KeyshareCommitmentRequest `json:"hw"`
+}
+
+type ProofPCommitmentMapV2 struct {
+	Commitments map[PublicKeyIdentifier]*big.Int `json:"c"`
+}
+
+func (cm *ProofPCommitmentMapV2) MarshalJSON() ([]byte, error) {
+	var encCM struct {
+		Commitments map[string]*big.Int `json:"c"`
+	}
+	encCM.Commitments = make(map[string]*big.Int)
+
+	for pki, c := range cm.Commitments {
+		pkiBytes, err := pki.MarshalText()
+		if err != nil {
+			return nil, err
+		}
+		encCM.Commitments[string(pkiBytes)] = c
+	}
+
+	return json.Marshal(encCM)
 }
 
 //